+++ lib/features/steps/value_steps.rb
@@ -39,6 +39,36 @@ Then('the {word} payload field {string} does not equal the stored value {string}
   assert_false(result.equal?, "Payload value: #{payload_value} equals stored value: #{stored_value}")
 end
 
+# Tests whether a payload field matches a previously stored payload value, ignoring case
+#
+# @step_input request_type [String] The type of request (error, session, build, etc)
+# @step_input field [String] The payload field to test
+# @step_input key [String] The key indicating a previously stored value
+Then('the {word} payload field {string} equals the stored value {string} ignoring case') do |request_type, field, key|
+  list = Maze::Server.list_for request_type
+  payload_value = Maze::Helper.read_key_path(list.current[:body], field)
+  stored_value = Maze::Store.values[key]
+  payload_value.downcase!
+  stored_value.downcase!
+  result = Maze::Compare.value(payload_value, stored_value)
+  assert_true(result.equal?, "Payload value: #{payload_value} does not equal stored value: #{stored_value}")
+end
+
+# Tests whether a payload field is distinct from a previously stored payload value, ignoring case
+#
+# @step_input request_type [String] The type of request (error, session, build, etc)
+# @step_input field [String] The payload field to test
+# @step_input key [String] The key indicating a previously stored value
+Then('the {word} payload field {string} does not equal the stored value {string} ignoring case') do |request_type, field, key|
+  list = Maze::Server.list_for request_type
+  payload_value = Maze::Helper.read_key_path(list.current[:body], field)
+  stored_value = Maze::Store.values[key]
+  payload_value.downcase!
+  stored_value.downcase!
+  result = Maze::Compare.value(payload_value, stored_value)
+  assert_false(result.equal?, "Payload value: #{payload_value} equals stored value: #{stored_value}")
+end
+
 # Tests whether a payload field is a number (Numeric according to Ruby)
 #
 # @step_input request_type [String] The type of request (error, session, build, etc)